+++ Editor/ValidationSuite/VettingContext/ValidationException.cs
@@ -1,10 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
+using UnityEngine.Serialization;
 
 namespace UnityEditor.PackageManager.ValidationSuite
 {
     [Serializable]
-    internal class ValidationException
+    class ValidationException
     {
         /// <summary>
         /// Name of validaiton test in which the exception is requested.
@@ -14,7 +15,7 @@ namespace UnityEditor.PackageManager.ValidationSuite
         /// <summary>
         /// Error for which the exception is requested.
         /// </summary>
-        public string ExceptionError;
+        public string ExceptionMessage;
 
         /// <summary>
         /// Package Version
@@ -23,8 +24,9 @@ namespace UnityEditor.PackageManager.ValidationSuite
     }
 
     [Serializable]
-    internal class ValidationExceptions
+    class ValidationExceptions
     {
-        public ValidationException[] Exceptions;
+        public ValidationException[] ErrorExceptions;
+        public ValidationException[] WarningExceptions;
     }
 }